==> src/04-his_orw_puller.Rmd <==
---
title: "his_orw_puller"
author: "Katie Willi"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(rvest)
library(jsonlite)
library(sf)
sf::sf_use_s2(FALSE)
```

First download all HIS Acces Databases from NPS IRMA:
```{r}
getHIS <- function(max = 1000, path = "data/in/HIS/"){
  
  #set base URL
  call <- "https://irmaservices.nps.gov/datastore/v4/rest"
  
  # Search for "Hydrographic" to get all HIS datasets
  dat <- httr::GET(paste0(call, "/QuickSearch?q=", "Hydrographic", "&top=", max)) 
  
  # convert content to text
  dat_text <- httr::content(dat, "text", encoding = "UTF-8")
  
  # parse data in JSON
  dat_json <- jsonlite::fromJSON(dat_text, flatten = TRUE)
  
  # convert items to data.frame
  dat_df <- dplyr::as_tibble(dat_json$items)
  
  # ID geospatial HIS databases to download
  dat_df_clean <- dat_df %>% 
    dplyr::filter(referenceType %in% c("Geospatial Dataset"),
                  grepl("Hydrographic and Impairment Statistic", title, ignore.case = TRUE)) %>% 
    dplyr::select(-newestVersion)
  
  # create empty vector to fill in downloadLink and name of download file
  dlLink <- vector("character", length = nrow(dat_df_clean))
  refName <- vector("character", length = nrow(dat_df_clean))
  
  #now get downloadLink (dataset download ID) for each item
  for (j in 1:nrow(dat_df_clean)){
    
    refID <- as.character(dat_df_clean[j, "referenceId"])
    
    res <- httr::GET(paste0(call, "/Reference/", refID, "/DigitalFiles"))
    
    #extract downloadLink
    resContent1 <- httr::content(res)[[1]]
    try(resContent2 <- httr::content(res)[[2]])
    
    if(grepl("xml", resContent1$fileName)==TRUE) {
      
      resContent <- resContent2
      
    } else {
      
      
      resContent <- resContent1
      
    }
    
    # if no file, no downloadLink so assign NA
    if(length(resContent) == 0){
      
      dlLink[j] <- NA
      
    } else {
      
      dlLink[j] <- resContent$downloadLink
      refName[j] <- resContent$fileName
      
    }
  }
  
  final_df <- dat_df_clean %>% 
    dplyr::mutate(downloadLink = dlLink,
                  referenceName = refName,
                  UNIT_CODE = str_sub(title,-4,-1)) %>%
    filter(!is.na(referenceName))
  
  for(i in 475:nrow(final_df)){
    
    df <- final_df[i,]
    
    # download the
    download.file(df$downloadLink, destfile = paste0(path, df$referenceName), method = 'curl')
    
    print(paste0(df[,11], " finished!"))
    
  }  
  
  
}
```

Then, munge each dataset's ORW table into a single file:
```{r}
path = "data/in/HIS/"

list.files(path = path, pattern = "*.zip") %>%
  map(~ unzip(zipfile = paste0(path, .), exdir = path, overwrite = TRUE)) 

list <- list.files(path = path, pattern = "*.mdb", full.names = TRUE)
small_list <- list.files(path = path, pattern = "*.mdb", full.names = FALSE)

converter <- function(list, small_list){
  
  db <- list
  con2 <- RODBC::odbcConnectAccess2007(db)
  RODBC::sqlTables(con2, tableType = "TABLE")$TABLE_NAME
  
  try(tblORW <- RODBC::sqlFetch(con2, "tblORW") %>%
        write_csv(paste0("data/mid/his_park_orw/", str_sub(small_list, 1, 4), '_tblORW.csv')), silent = TRUE)
  
  ET_Park <- RODBC::sqlFetch(con2, "ET_Park") 
  
  try(ET_Park <- ET_Park %>% left_join(tblORW, by = "ORW_ID"), silent = TRUE)
  
  ET_Park <- ET_Park %>% 
    mutate(UNIT_CODE = str_sub(small_list, 1, 4)) %>%
    write_csv(paste0("data/mid/his_park_orw/", str_sub(small_list, 1, 4), '_ET_Park_plus.csv'))
  
  print(paste0(str_sub(small_list, 1, 4), " HIS data converted to .csv"))
}

map2(list, small_list, possibly(converter, otherwise = print(" has no HIS data!")))

perm_id <- function(x){
  x <- read.csv(x) %>%
    select(5) %>%
    rename(permanent_identifier = 1) %>%
    mutate(permanent_identifier = as.character(permanent_identifier))
  
  return(x)
  
}

perm_id <- list.files(path = "data/mid/his_park_orw", pattern = "*ET_Park_plus.csv", full.names = TRUE) %>%
  map(~perm_id(.)) %>% 
  bind_rows()

cleaner <- function(x){
  
  x <- read.csv(x)
  
  names(x) <- tolower(names(x))
  
  x <- x %>%
    dplyr::select(any_of(c("unit_code", "gnis_id", "entity_id", "gnis_name",
                           "orw_id", "state", "reachcode",
                           "designation_level", "designation_name", 
                           "entire_extent_of_orw")))
  
  return(x)
  
}


orw_only <- list.files(path = "data/mid/his_park_orw", pattern = "*ET_Park_plus.csv", full.names = TRUE) %>%
  map(~cleaner(.)) %>%
  bind_rows() %>%
  cbind(perm_id) %>%
  filter(!is.na(orw_id)) %>%
  mutate(Tier = ifelse(grepl("Outstanding National Resource Water|ONRW|Tier 3|3", designation_name, ignore.case = TRUE), "Tier 3", "Tier 2.5")) %>%
  mutate(Tier = ifelse(is.na(Tier) & unit_code == "BISO", "Tier 3",
                       ifelse(is.na(Tier) & unit_code == "GRSM", "Tier 2.5", Tier))) %>%
  mutate(ReachCode = as.character(reachcode))

saveRDS(orw_only, 'data/mid/parks_with_orw.RDS')
```
