==> src/05-attains_by_parks.Rmd <==
---
title: "05 - Statistics by park unit"
author: "Kathryn Willi"
date: "2023-05-16"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(sf)
library(data.table)
```

This is where we develop ALL statistics associated with the Delivery 1 report. This includes: 

- How many parks have an impaired assessment unit?
- What are the top impairments across park units?
- Which parks have water, but no assessment unit (i.e. which parks are unassessed)?
- Which parks have outstanding resource waters?

All statistics are only run on parks with water. We identified which parks had water by finding at park boundary unit that did not have any intersecting NHDPlus HR water body, area, of flowline features. We also utilized a table from the Hydrographic and Impairment Statistics database listing parks they consider to have to hydrography.

#### Getting the data

Load in necessary data sets:
```{r}
map(list.files('src/fxn', full.names = TRUE), source)

# Load in park boundaries:
parks <- st_read("data/in/nps_boundary.shp") 

# Load in NPCA regions
npca_regions <- st_read("data/in/npca_regions.shp") 

# Load in all geospatial ATTAINS data:
attains_areas <- readRDS('data/mid/attains_au_areas.RDS')

attains_lines <- readRDS('data/mid/attains_au_lines.RDS')

attains_points <- readRDS('data/mid/attains_au_points.RDS')

# Load in ATTAINS water quality standard parameter data:
attains_assmnt_parms <- readRDS("data/mid/attains_au_assmnt_parms.RDS")

```

Subset and clip the ATTAINS geospatial data to only include assessment units within NPS Park Boundaries.
```{r}
nps_attains_areas_raw <- attains_areas %>% 
  subsetATTAINS(layer = ., sub_by = parks, type = "area") %>%
  mutate(assessment_type = "AREA",
         combo = paste0(UNIT_CODE, "-", assessmentunitidentifier))

coastal_waters <- st_read('data/bad/erase_temp.shp') %>% select(UNIT_CODE) %>% mutate(CATCH_TYPE = "COASTAL", nhdplusid = NA, SHAPE = geometry) %>% sf::st_transform(3857) 

nps_attains_areas_near_wb <- nps_attains_areas_raw %>%
  # Only select area features within 10 meters of an NHD water feature.
  # I have found that "area" ATTAINS features are the most coarse resolution
  # of all ATTAINS features. Therefore, this step is to verify that ATTAINS
  # areas within park units are actually linked to an existing water feature
  # in the park:
  st_buffer(10) %>%
  # This layer I am joining by is the clipped NHD-HR features within park units.
  # Meaning, only keep ATTAINS area features that are near on directly contain
  # and NHD HR water feature:
  st_join(., readRDS('data/mid/nhdhr_park_waterbody.RDS'), left = FALSE) %>% 
  select(combo)

nps_attains_areas_near_flowlines <- nps_attains_areas_raw %>%
  # Only select area features within 10 meters of a water feature
  st_buffer(10) %>%
  st_join(., readRDS('data/mid/nhdhr_park_flow.RDS'), left = FALSE) %>%
  select(combo)

nps_attains_areas_near_areas <- nps_attains_areas_raw %>%
  # Only select area features within 10 meters of a water feature
  st_buffer(10) %>%
  st_join(., readRDS('data/mid/nhdhr_park_area.RDS'), left = FALSE) %>%
  select(combo)

nps_attains_areas_near_coasts <- nps_attains_areas_raw %>%
  # Only select area features within 10 meters of a water feature
  st_buffer(10) %>%
  st_join(., select(coastal_waters, CATCH_TYPE), left = FALSE) %>%
  select(combo)

nps_attains_areas <- nps_attains_areas_raw %>%
  filter(combo %in% c(nps_attains_areas_near_wb$combo, nps_attains_areas_near_flowlines$combo, 
                      nps_attains_areas_near_areas$combo, nps_attains_areas_near_coasts$combo)| 
           # Cannot do this exercise in states without NHD:
           STATE %in% c("AK","HI","PR","VI","AS","GU","MP"))

saveRDS(nps_attains_areas,'data/mid/nps_attains_park_area_right.RDS')

nps_attains_lines <- attains_lines %>%
  subsetATTAINS(layer = ., sub_by = parks, type = "line") %>%
  mutate(assessment_type = "LINE")
saveRDS(nps_attains_lines,'data/mid/nps_attains_park_line_right.RDS')

nps_attains_points <- attains_points %>%
  subsetATTAINS(layer = ., sub_by = parks, type = "point") %>%
  mutate(assessment_type = "POINT")
saveRDS(nps_attains_points,'data/mid/nps_attains_park_point_right.RDS')

# this represents all assessment units (area, point, linear) that are physically within the park boundaries
nps_all_attains <- bind_rows(as_tibble(nps_attains_areas), as_tibble(nps_attains_lines), as_tibble(nps_attains_points)) %>%
  as_tibble(.) %>%
  # join to assessment unit parameter data:
  dplyr::left_join(.,
                   # remove duplicate/unnecessary columns:
                   dplyr::select(attains_assmnt_parms, -c(assmnt_joinkey, organizationid, submissionid, orgtype, tas303d, reportingcycle, waterbodyreportlink, GLOBALID)),
                   by = "assessmentunitidentifier", multiple = "all")

raw_parks <- st_read("data/in/nps_boundary.shp") 
# Lastly remove all data linked to parks without water features in them (CONTUS only):
no_water <- raw_parks %>%
  # cannot include these states in this exercise as they don't have nhd hr in them to begin with
  # but, all parks outside CONUS have water (based on a manual inspection of them)
  filter(!STATE %in% c("AK","HI","PR","VI","AS","GU","MP")) %>%
  # this is a weird lil park in the Middle of Lake Erie:
  filter(UNIT_CODE != "PEVI") %>%
  # clipped water features from the NHD HR within park units:
  filter(!UNIT_CODE %in% readRDS('data/mid/nhdhr_park_area.RDS')$UNIT_CODE) %>%
  filter(!UNIT_CODE %in% readRDS('data/mid/nhdhr_park_flow.RDS')$UNIT_CODE) %>%
  filter(!UNIT_CODE %in% readRDS('data/mid/nhdhr_park_waterbody.RDS')$UNIT_CODE) %>%
  # table developed by NPS (but edited by me to include more parks that have no water - based on
  # a manual inspection) that lists parks without water:
  bind_rows(filter(raw_parks, UNIT_NAME %in% read_csv('data/in/no_water_parks.csv')$UNIT_NAME)) %>%
  distinct(UNIT_NAME, .keep_all = TRUE) %>%
  mutate(hydro = "No Water Features") %>%
  st_drop_geometry()

nps_all_attains <- nps_all_attains %>%
  filter(!UNIT_CODE %in% no_water$UNIT_CODE) %>%
  bind_rows(no_water) 

saveRDS(nps_all_attains, 'data/out/nps_all_wq_data_final_right.RDS')
```

#### Exploring the data
```{r}
nps_all_attains <- readRDS('data/out/nps_all_wq_data_final_right.RDS')

colors <- data.frame(
  overallstatus = c("Not Supporting", "Fully Supporting", "Not Assessed", NA),
  col = c("#DC851E", "#059FA4", "#A1A522", "#C2CAD7"),
  Priority = c(1, 2, 3, 4)
)

# making the USA "pretty":
prettier <- USA.state.boundaries::state_boundaries_wgs84 %>% filter(NAME %in% c("Hawaii", "Michigan") & TYPE == "Land") %>%
  st_transform(3857) %>%
  mutate(geometry = Shape)
states <- tigris::states(resolution = "5m") %>% filter(!NAME %in% c("Michigan", "Hawaii")) %>%
  st_transform(3857) %>%
  bind_rows(prettier) %>%
  usa_mapper()
rm(prettier)

parks_mapper <- parks %>%
  st_centroid() %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  distinct(UNIT_CODE, .keep_all = TRUE) %>%
  usa_mapper()

parks <- parks %>%
  st_join(., npca_regions)
```

##### Impaired Parks
```{r}
nps_all_attains <- nps_all_attains %>%
  left_join(colors, by = "overallstatus") %>%
  dplyr::filter(is.na(hydro))

# how many parks have impairments?
count <- nps_all_attains %>% 
  dplyr::filter(ircategory %in% c("5A","5","4C","4B", "4A")) %>%
  distinct(UNIT_CODE) %>% nrow()

total_w_water <- parks %>%
  distinct(UNIT_CODE) %>%
  filter(!UNIT_CODE %in% no_water$UNIT_CODE)

# how many with impairments on the 303d list?
nps_all_attains %>% filter(parametercategorycode %in% c("5A","5")) %>%
  distinct(UNIT_CODE, .keep_all = TRUE) %>% nrow()

# where are the impaired parks?
impaired_parks <- nps_all_attains %>%
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B", "4A")) %>% 
  distinct(UNIT_CODE, .keep_all = TRUE)

impaired_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% impaired_parks$UNIT_CODE) %>%
  st_centroid() %>%
  distinct(UNIT_CODE, .keep_all = TRUE) %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = impaired_mapper, fill ="#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/impaired_parks.jpg')#, width = 12.14, height = 7.5, units = "cm")
```

##### Common impairments
```{r}
# What impairments are the most common across park units?
top_impairments <- nps_all_attains %>% 
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B","4A")) %>%
  distinct(UNIT_CODE, parametergroupname) %>%
  group_by(parametergroupname) %>%
  summarize(count = n())

pathogens <- nps_all_attains %>%
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B", "4A")) %>% 
  dplyr::filter(parametergroupname == "PATHOGENS") %>%
  distinct(UNIT_CODE, .keep_all =TRUE)
pathogen_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% pathogens$UNIT_CODE) %>%
  st_centroid() %>%
  distinct(UNIT_CODE, .keep_all =TRUE) %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

mercury <- nps_all_attains %>%
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B", "4A")) %>% 
  dplyr::filter(parametergroupname == "MERCURY") %>%
  distinct(UNIT_CODE, .keep_all =TRUE)
mercury_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% mercury$UNIT_CODE) %>%
  st_centroid() %>%
  distinct(UNIT_CODE, .keep_all =TRUE) %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

metals <- nps_all_attains %>%
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B", "4A")) %>% 
  dplyr::filter(parametergroupname == "METALS (OTHER THAN MERCURY)") %>%
  distinct(UNIT_CODE, .keep_all =TRUE)
metals_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% metals$UNIT_CODE) %>%
  st_centroid() %>%
  distinct(UNIT_CODE, .keep_all =TRUE) %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

nutrients <- nps_all_attains %>%
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B", "4A")) %>% 
  dplyr::filter(parametergroupname == "NUTRIENTS") %>%
  distinct(UNIT_CODE, .keep_all =TRUE)
nutrients_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% nutrients$UNIT_CODE) %>%
  st_centroid() %>%
  distinct(UNIT_CODE, .keep_all =TRUE) %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

do <- nps_all_attains %>%
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B", "4A")) %>% 
  dplyr::filter(parametergroupname == "ORGANIC ENRICHMENT/OXYGEN DEPLETION") %>%
  distinct(UNIT_CODE, .keep_all =TRUE)
do_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% do$UNIT_CODE) %>%
  st_centroid() %>%
  distinct(UNIT_CODE, .keep_all =TRUE) %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

pcbs <- nps_all_attains %>%
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B", "4A")) %>% 
  dplyr::filter(parametergroupname == "POLYCHLORINATED BIPHENYLS (PCBS)") %>%
  distinct(UNIT_CODE, .keep_all =TRUE)
pcbs_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% pcbs$UNIT_CODE) %>%
  st_centroid() %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = pathogen_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/pathogens.jpg')

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = mercury_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/mercury.jpg')

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = metals_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/metals.jpg')

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = nutrients_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/nutrients.jpg')

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = do_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/do.jpg')

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = pcbs_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/pcbs.jpg')
```

##### Impairments by NPCA Region
```{r}
# What impairments are the most common across regions?
top_npca_impairments <- nps_all_attains %>% 
  dplyr::filter(parametercategorycode %in% c("5A","5","4C","4B","4A")) %>%
  distinct(UNIT_CODE, office, parametergroupname) %>%
  group_by(office, parametergroupname) %>%
  summarize(count = n()) %>%
  ungroup() %>%
  arrange(desc(count)) %>% 
  group_by(office) %>%
  slice(1:10)
```

##### Which parks have no water quality data?
```{r}
no_data <- parks %>% 
  anti_join(., nps_all_attains, by = "UNIT_CODE") %>%
  filter(!UNIT_CODE %in% no_water$UNIT_CODE) %>% 
  distinct(UNIT_CODE, .keep_all = TRUE) 
no_data_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% no_data$UNIT_CODE) %>%
  st_centroid() %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = no_data_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21) 
ggsave('data/out/figs/no_data.jpg')
```

##### Which parks have ORWs?
```{r}
park_table <- parks %>%
  st_drop_geometry() %>%
  distinct(UNIT_CODE)

any <- readRDS('data/mid/parks_with_orw.RDS') %>%
  inner_join(park_table, by = c("unit_code" = "UNIT_CODE")) %>%
  distinct(unit_code) %>%
  filter(!unit_code %in% no_water$UNIT_CODE)

none <- park_table %>%
  anti_join(., any, by = c("UNIT_CODE" = "unit_code")) %>%
  # don't include parks without water in the count:
  dplyr::filter(!UNIT_CODE %in% no_water$UNIT_CODE)

orw <- readRDS('data/mid/parks_with_orw.RDS') %>%
  inner_join(park_table, by = c("unit_code"="UNIT_CODE")) %>%
  distinct(unit_code, Tier) %>%
  dplyr::filter(Tier == "Tier 2.5")

onrw <- readRDS('data/mid/parks_with_orw.RDS') %>%
  inner_join(park_table, by = c("unit_code"="UNIT_CODE")) %>%
  distinct(unit_code,Tier) %>%
  dplyr::filter(Tier == "Tier 3")

both <- readRDS('data/mid/parks_with_orw.RDS') %>%
  inner_join(park_table, by = c("unit_code" = "UNIT_CODE")) %>%
  distinct(unit_code, Tier) %>%
  group_by(unit_code) %>%
  summarize(count = n()) %>%
  dplyr::filter(count > 1)

orw_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% orw$unit_code) %>%
  st_centroid() %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

onrw_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% onrw$unit_code) %>%
  st_centroid() %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

both_mapper <- parks %>%
  dplyr::filter(UNIT_CODE %in% both$unit_code) %>%
  st_centroid() %>%
  st_join(st_transform(tigris::states(), 3857)) %>%
  usa_mapper()

ggplot() +
  theme_void() +
  geom_sf(data = states, fill = NA) +
  geom_sf(data = parks_mapper, color = "#C2CAD7") + 
  geom_sf(data = orw_mapper, fill = "#A1A522", color = "black", size = 3, pch = 21) +
  geom_sf(data = onrw_mapper, fill = "#059FA4", color = "black", size = 3, pch = 21) +
  geom_sf(data = both_mapper, fill = "#DC851E", color = "black", size = 3, pch = 21)
ggsave('data/out/figs/orws.jpg')
```
