==> src/06-geo_grabber.Rmd <==
---
title: "park_geospatial_data"
author: "Katie Willi"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Produces a list of all relevant AOIs for querying water quality data portals for each park unit + 10 km radius around each park unit.

```{r}
library(tidyverse)
library(sf)
library(tigris)

options(tigris_use_cache = TRUE)
states <- tigris::states()
counties <- tigris::counties()
```

```{r}
parks <- st_read('data/in/nps_boundary.shp') %>%
  filter(!UNIT_CODE %in% c("APPA", "MAPR"))
parks_buf <- parks %>%
  # 10 km buffer:
  sf::st_buffer(dist = 10000)
park_list <- unique(parks$UNIT_CODE) 
# Remove the Appalachian Trail and MAPR from the list. They're too big.
park_list <- park_list[! park_list %in% c("MAPR","APPA")]

geo_grabber <- function(park_list){
  
  parks <- filter(parks_buf, UNIT_CODE %in% park_list) %>%
    st_transform(4269) %>%
    group_by(UNIT_CODE) %>%      
    summarize()
  
  state <- states %>%
    .[parks,] 
  
  county <- counties %>%
    .[parks,] %>%
    select(NAMELSAD, STATEFP) %>% 
    left_join(st_drop_geometry(states), by = "STATEFP") %>%
    select(NAME, NAMELSAD, STATEFP) %>% 
    mutate(NAME = paste0(NAMELSAD, ", ", NAME),
           TYPE = "COUNTY") %>%
    select(NAME, TYPE) %>%
    st_drop_geometry()
  
  places <- tigris::places(state = unique(state$STUSPS), filter_by = parks) %>% 
    select(NAMELSAD, STATEFP) %>% 
    left_join(st_drop_geometry(states), by = "STATEFP") %>%
    select(NAME, NAMELSAD, STATEFP) %>% 
    mutate(NAME = paste0(NAMELSAD, ", ", NAME),
           TYPE = "PLACE") %>%
    select(NAME, TYPE) %>%
    st_drop_geometry()
  
  huc12 <- nhdplusTools::get_huc(AOI = parks, type = "huc12") %>% 
    mutate(TYPE = "HUC12") %>%
    select(NAME = huc12,
           TYPE) %>%
    st_drop_geometry()
  
  zipcode <- tigris::zctas(filter_by = parks) %>% 
    select(NAME = 1) %>%
    mutate(TYPE = "ZIPCODE") %>%
    st_drop_geometry()
  
  bbox <- st_bbox(parks) %>% as_tibble() %>%
    mutate(TYPE = c("bbox:xmin", "bbox:ymin", "bbox:xmax", "bbox:ymax")) %>%
    rename(NAME = x) %>%
    mutate(NAME = as.character(NAME)) %>%
    st_drop_geometry()
  
  print(paste0(parks$UNIT_CODE, " done!"))
  
  me <- rbind(places, county, huc12, zipcode, bbox) %>%
    mutate(UNIT_CODE = unique(parks$UNIT_CODE)) 
  # back up
  write_csv(me, paste0('data/mid/aoi/', parks$UNIT_CODE, '.csv'))
  return(me)
  
}

park_list %>%
  map(~geo_grabber(.)) %>%
  bind_rows() 
```

Two parks (Appalachian Trail and the Manhattan Project) require a different approach for pulling this AOI info. Here, I am splitting the parks up into state "chunks", then performing the same function on them. 
```{r}
sf_use_s2(FALSE)

baddie_grabber <- function(baddies_list){
  
  parks <- filter(baddies_buf, rowid == baddies_list) %>%
    select(rowid, UNIT_CODE) %>%
    st_transform(4269) #%>%
    # group_by(UNIT_CODE) %>%      
    # summarize()
  
  state <- states %>%
    .[parks,] 
  
  county <- counties %>%
    .[parks,] %>%
    select(NAMELSAD, STATEFP) %>% 
    left_join(st_drop_geometry(states), by = "STATEFP") %>%
    select(NAME, NAMELSAD, STATEFP) %>% 
    mutate(NAME = paste0(NAMELSAD, ", ", NAME),
           TYPE = "COUNTY") %>%
    select(NAME, TYPE) %>%
    st_drop_geometry()
  
  places <- tigris::places(state = unique(state$STUSPS), filter_by = parks) %>% 
    select(NAMELSAD, STATEFP) %>% 
    left_join(st_drop_geometry(states), by = "STATEFP") %>%
    select(NAME, NAMELSAD, STATEFP) %>% 
    mutate(NAME = paste0(NAMELSAD, ", ", NAME),
           TYPE = "PLACE") %>%
    select(NAME, TYPE) %>%
    st_drop_geometry()
  
  huc12 <- nhdplusTools::get_huc(AOI = parks, type = "huc12") %>% 
    mutate(TYPE = "HUC12") %>%
    select(NAME = huc12,
           TYPE) %>%
    st_drop_geometry()
  
  zipcode <- tigris::zctas(filter_by = parks) %>% 
    select(NAME = 1) %>%
    mutate(TYPE = "ZIPCODE") %>%
    st_drop_geometry()
  
  print(paste0(parks$UNIT_CODE, " done!"))
  
  me <- rbind(places, county, huc12, zipcode) %>%
    mutate(UNIT_CODE = unique(parks$UNIT_CODE)) 
  # back up
  write_csv(me, paste0('data/mid/aoi/', parks$UNIT_CODE, "-", parks$rowid, '.csv'))
  return(me)
  
}

baddies_buf <- st_read('data/in/nps_boundary.shp') %>%
  filter(UNIT_CODE %in% c("APPA", "MAPR")) %>%
  st_buffer(dist = 10000) %>%
  st_transform(4269) %>%
  sf::st_intersection(states) %>%
  rowid_to_column() %>%
  select(rowid, UNIT_CODE)

baddies_list <- baddies_buf$rowid

bad_sites <- baddies_list %>%
  map(~baddie_grabber(.)) %>%
  bind_rows() #%>%


bbox <- st_bbox(filter(parks_buf, UNIT_CODE == "APPA")) %>% as_tibble() %>%
  mutate(TYPE = c("bbox:xmin", "bbox:ymin", "bbox:xmax", "bbox:ymax"))# %>%
  rename(NAME = x) %>%
  mutate(UNIT_CODE="APPA") %>%
  st_drop_geometry()
```

Bind all AOIs by parks into a single file:
```{r}
all_aois <- list.files('data/mid/aoi/', full.names = TRUE) %>%
  map_dfr(~read_csv(.)) %>%
  distinct(.keep_all = TRUE) %>%
  #Bounding boxes too big and not useful for these parks so I'm not including them:
  bind_rows(tibble(UNIT_CODE = c("APPA", "MAPR"),
                   NAME = c(NA, NA),
                   TYPE = c("bbox", "bbox"))) 

write_csv(all_aois, 'data/out/D2_Appendix_A2.csv')
```
